Create shared folder; add createSnippet in SnippetService
