+++ front/src/app/snippets/shared/snippet.ts
