+++ front/src/app/snippets/snippets.module.ts
@@ -1,6 +1,6 @@
 import { SharedModule } from '../shared/shared.module';
 import { SnippetsRoutingModule } from './snippets-routing.module';
-import { SnippetService } from './snippet/snippet.service';
+import { SnippetService } from './shared/snippet.service';
 import { SnippetComponent } from './snippet/snippet.component';
 import { NgModule } from '@angular/core';
 import { SnippetFormComponent } from './snippet-form/snippet-form.component';
